Dry out pruner methods
