+++ misk-rate-limiting-bucket4j-dynamodb-v1/api/misk-rate-limiting-bucket4j-dynamodb-v1.api
@@ -6,11 +6,12 @@ public final class misk/ratelimiting/bucket4j/dynamodb/v1/DynamoDbV1Bucket4jRate
 	public final fun providedRateLimiter (Ljava/time/Clock;Lcom/amazonaws/services/dynamodbv2/AmazonDynamoDB;Lio/micrometer/core/instrument/MeterRegistry;)Lwisp/ratelimiting/RateLimiter;
 }
 
-public final class misk/ratelimiting/bucket4j/dynamodb/v1/DynamoDbV1BucketPruner : wisp/ratelimiting/RateLimitPruner {
+public final class misk/ratelimiting/bucket4j/dynamodb/v1/DynamoDbV1BucketPruner : wisp/ratelimiting/bucket4j/Bucket4jPruner {
 	public static final field Companion Lmisk/ratelimiting/bucket4j/dynamodb/v1/DynamoDbV1BucketPruner$Companion;
 	public fun <init> (Ljava/time/Clock;Lcom/amazonaws/services/dynamodbv2/AmazonDynamoDB;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;)V
 	public fun <init> (Ljava/time/Clock;Lcom/amazonaws/services/dynamodbv2/AmazonDynamoDB;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;I)V
 	public synthetic fun <init> (Ljava/time/Clock;Lcom/amazonaws/services/dynamodbv2/AmazonDynamoDB;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;IILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun getClockTimeMeter ()Lwisp/ratelimiting/bucket4j/ClockTimeMeter;
 	public fun prune ()V
 }
