+++ misk-rate-limiting-bucket4j-mysql/api/misk-rate-limiting-bucket4j-mysql.api
@@ -6,12 +6,13 @@ public final class misk/ratelimiting/bucket4j/mysql/MySQLBucket4jRateLimiterModu
 	public final fun providedRateLimiter (Ljava/time/Clock;Lcom/google/inject/Injector;Lio/micrometer/core/instrument/MeterRegistry;)Lwisp/ratelimiting/RateLimiter;
 }
 
-public final class misk/ratelimiting/bucket4j/mysql/MySQLBucketPruner : wisp/ratelimiting/RateLimitPruner {
+public final class misk/ratelimiting/bucket4j/mysql/MySQLBucketPruner : wisp/ratelimiting/bucket4j/Bucket4jPruner {
 	public static final field Companion Lmisk/ratelimiting/bucket4j/mysql/MySQLBucketPruner$Companion;
 	public fun <init> (Ljava/time/Clock;Ljavax/sql/DataSource;Ljava/lang/String;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;Ljava/lang/String;)V
 	public fun <init> (Ljava/time/Clock;Ljavax/sql/DataSource;Ljava/lang/String;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;Ljava/lang/String;Z)V
 	public fun <init> (Ljava/time/Clock;Ljavax/sql/DataSource;Ljava/lang/String;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;Ljava/lang/String;ZJ)V
 	public synthetic fun <init> (Ljava/time/Clock;Ljavax/sql/DataSource;Ljava/lang/String;Lio/micrometer/core/instrument/MeterRegistry;Ljava/lang/String;Ljava/lang/String;ZJILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun getClockTimeMeter ()Lwisp/ratelimiting/bucket4j/ClockTimeMeter;
 	public fun prune ()V
 }
