+++ wisp/wisp-rate-limiting/bucket4j/src/main/kotlin/wisp/ratelimiting/bucket4j/Bucket4jPruner.kt
@@ -1,0 +1,24 @@
+package wisp.ratelimiting.bucket4j
+
+import io.github.bucket4j.distributed.remote.RemoteBucketState
+import io.github.bucket4j.distributed.serialization.DataOutputSerializationAdapter
+import wisp.ratelimiting.RateLimitPruner
+import java.io.ByteArrayInputStream
+import java.io.DataInputStream
+
+abstract class Bucket4jPruner : RateLimitPruner {
+  abstract val clockTimeMeter: ClockTimeMeter
+
+  protected fun isBucketStale(state: RemoteBucketState): Boolean {
+    val refillTimeNanos = state.calculateFullRefillingTime(clockTimeMeter.currentTimeNanos())
+    return refillTimeNanos <= 0L
+  }
+
+  protected fun deserializeState(bytes: ByteArray): RemoteBucketState {
+    val inputStream = DataInputStream(ByteArrayInputStream(bytes))
+    return RemoteBucketState.SERIALIZATION_HANDLE.deserialize(
+      DataOutputSerializationAdapter.INSTANCE,
+      inputStream
+    )
+  }
+}
